--- src/hohland_cblp.ProductService.Domain/Entities/ProductType.cs
@@ -1,9 +1,0 @@
-namespace hohland_cblp.ProductService.Domain.Entities;
-
-public enum ProductType
-{
-    General, 
-    HouseholdChemicals,
-    Appliances, 
-    Food
-}
