--- src/hohland_cblp.ProductService.Domain/Validators/ProductValidator.cs
@@ -1,28 +1,0 @@
-﻿using FluentValidation;
-using hohland_cblp.ProductService.Domain.Entities;
-
-
-namespace hohland_cblp.ProductService.Domain.Validators;
-
-public class ProductValidator : AbstractValidator<Product>
-{
-    public ProductValidator()
-    {
-        RuleFor(item => item.Id)
-            .NotNull();
-        RuleFor(item => item.Name)
-            .NotEmpty()
-            .MinimumLength(3)
-            .MaximumLength(20);
-        RuleFor(item => item.Price)
-            .NotNull();
-        RuleFor(item => item.Weight)
-            .NotNull();
-        RuleFor(item => item.WarehouseId)
-            .NotNull();
-        RuleFor(item => item.CreationDate)
-            .NotNull();
-        RuleFor(item => item.ProductType)
-            .NotNull();
-    }
-}
