--- src/hohland_cblp.ProductService.Infrastructure/hohland_cblp.ProductService.Infrastructure.csproj
@@ -1,30 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <RootNamespace>hohland_cblp.ProductService.Infrastructure</RootNamespace>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\hohland_cblp.ProductService.Domain\hohland_cblp.ProductService.Domain.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <PackageReference Include="AutoMapper" Version="12.0.1" />
-        <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1" />
-        <PackageReference Include="Google.Api.CommonProtos" Version="2.10.0" />
-        <PackageReference Include="Google.Protobuf" Version="3.24.3" />
-        <PackageReference Include="Grpc.AspNetCore" Version="2.57.0" />
-        <PackageReference Include="Grpc.AspNetCore.Server" Version="2.57.0" />
-        <PackageReference Include="Grpc.Core" Version="2.46.6" />
-        <PackageReference Include="Grpc.Core.Api" Version="2.57.0" />
-        <PackageReference Include="Grpc.Tools" Version="2.58.0">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-        <Protobuf Include="**\*.proto" Exclude="google\**" />
-    </ItemGroup>
-    
-</Project>
