+++ src/hohland_cblp.ShopBackend.Domain/Entities/ProductType.cs
@@ -1,0 +1,9 @@
+namespace hohland_cblp.ShopBackend.Domain.Entities;
+
+public enum ProductType
+{
+    General, 
+    HouseholdChemicals,
+    Appliances, 
+    Food
+}
