+++ src/hohland_cblp.ShopBackend.Infrastructure/DataBaseEntities/ProductEntityV1.cs
@@ -1,0 +1,6 @@
+﻿namespace hohland_cblp.ShopBackend.Infrastructure.DataBaseEntities;
+
+public class ProductEntityV1
+{
+    
+}
