+++ src/hohland_cblp.ShopBackend.Infrastructure/PostgresInfrastructure/Postgres.cs
@@ -1,0 +1,37 @@
+using FluentMigrator.Runner;
+using hohland_cblp.ShopBackend.Infrastructure.DataBaseEntities;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
+using Npgsql;
+using Npgsql.NameTranslation;
+
+namespace hohland_cblp.ShopBackend.Infrastructure.PostgresInfrustructure;
+
+public static class Postgres
+{
+    private static readonly INpgsqlNameTranslator Translator = new NpgsqlSnakeCaseNameTranslator();
+
+    /// <summary>
+    /// Map DAL models to composite types (enables UNNEST)
+    /// </summary>
+    public static void MapCompositeTypes()
+    {
+        var mapper = NpgsqlConnection.GlobalTypeMapper;
+        Dapper.DefaultTypeMap.MatchNamesWithUnderscores = true;
+        
+        mapper.MapComposite<ProductEntityV1>("user_v1", Translator);
+    }
+
+    /// <summary>
+    /// Add migration infrastructure
+    /// </summary>
+    public static void AddMigrations(IServiceCollection services, string connectionString)
+    {
+        services.AddFluentMigratorCore()
+            .ConfigureRunner(rb => rb.AddPostgres()
+                .WithGlobalConnectionString(_ => connectionString)
+                .ScanIn(typeof(Postgres).Assembly).For.Migrations()
+            )
+            .AddLogging(lb => lb.AddFluentMigratorConsole());
+    }
+}
