+++ src/hohland_cblp.ShopBackend.Infrastructure/Proto/product.proto
