+++ src/hohland_cblp.ShopBackend.Infrastructure/google/api/annotations.proto
