+++ src/hohland_cblp.ShopBackend.Infrastructure/google/api/http.proto
